+++ extension/healthcheckextension/config.go
@@ -16,6 +16,7 @@ package healthcheckextension
 
 import (
 	"go.opentelemetry.io/collector/config/configmodels"
+	"go.opentelemetry.io/collector/config/confignet"
 )
 
 // Config has the configuration for the extension enabling the health check
@@ -25,5 +26,11 @@ type Config struct {
 
 	// Port is the port used to publish the health check status.
 	// The default value is 13133.
+	// Deprecated: use Endpoint instead.
 	Port uint16 `mapstructure:"port"`
+
+	// TCPAddr represents a tcp endpoint address that is to publish the health
+	// check status.
+	// The default endpoint is "localhost:13133".
+	TCPAddr confignet.TCPAddr `mapstructure:",squash"`
 }